+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Commit messages follow [Conventional Commits](https://www.conventionalcommits.or
 ### Added
 
 - Article title and body text can be selected and copied
+- Tapping an article image opens a zoomable gallery
 
 ## [0.0.1] - 2026-08-14